ripple directive moved inside directives/ripple folder
